17.text_to_speech/requirements.txt
pytorch
torchaudio
deep-phonemizer